+++ examples/button.rs
@@ -2,24 +2,26 @@ use win_toast_notify::{WinToastNotify, Action, ActivationType};
 
 fn main() {
     WinToastNotify::new()
-        .set_title("Hellow World")
-        .set_messages(vec![
-            "There are two buttons here"
-        ])
         .set_actions(vec![
             Action {
                 activation_type: ActivationType::Protocol,
-                action_content: "Open Github",
-                arguments: "https://github.com/",
-                image_url: None,
+                action_content: "Url".to_string(),
+                arguments: "https://www.google.com/".to_string(),
+                image_url: None
             },
             Action {
                 activation_type: ActivationType::Protocol,
-                action_content: "Open Wallpaper",
-                arguments: r"C:\Windows\Web\Screen\img101.jpg",
-                image_url: None,
-            }
-        ])
-        .show()
-        .expect("Failed to show toast notification")
+                 action_content: "File".to_string(),
+                 arguments: r"C:\Windows\Web\Screen\img104.jpg".to_string(),
+                 image_url: None
+             },
+            Action {
+                 activation_type: ActivationType::Protocol,
+                 action_content: "Folder".to_string(),
+                 arguments: r"$env:USERPROFILE\Downloads".to_string(),   // PowerShell supports using environment variables
+                 image_url: None
+             }
+         ])
+         .show()
+         .expect("Failed to show toast notification");
 }