+++ examples/progress_bat_1.rs
@@ -1,4 +1,4 @@
-use win_toast_notify::{WinToastNotify, Duration, Progress};
+use win_toast_notify::{WinToastNotify, Duration};
 
 fn main() {
     let tag = "weekly-playlist";
@@ -10,7 +10,7 @@ fn main() {
     WinToastNotify::new()
         .set_duration(Duration::Long)
         .set_title("Downloading your weekly playlist...")
-        .set_progress(Progress {tag, title, status, value, value_string})
+        .set_progress(tag, title, &status, value, &value_string)
         .show()
         .expect("Failed to show toast notification");
 
@@ -19,11 +19,11 @@ fn main() {
         value = i as f32 / 10.0;
         if i != 10 {
             value_string = format!("{}/10 songs", i);
-            WinToastNotify::progress_update(None, tag, value, value_string).expect("Failed to update");
+            WinToastNotify::progress_update(None, tag, value, &value_string).expect("Failed to update");
         } else {
             status = String::from("Completed");
             value_string = String::from("10/10 songs");
-            WinToastNotify::progress_complete(None, tag, status, value_string).expect("Failed to complete");
+            WinToastNotify::progress_complete(None, tag, &status, &value_string).expect("Failed to complete");
         };
     };
 }